Create Azure_Find_Expired_Token.ps1
Make sure you have the AzureAD module installed before running the script. The script connects to Azure AD using `Connect-AzureAD` to establish a session.

It retrieves the list of applications using `Get-AzureADApplication` and then iterates through each application using a foreach loop.

For each application, it retrieves the access token using `Get-AzureADServicePrincipalCredential` and the application's ObjectId. This token is used for authentication when making requests to the Microsoft Graph API.

The script sends a GET request to the tokenLifetimePolicies endpoint of the Graph API for each application. It retrieves the list of token lifetime policies associated with the application.

The script filters the policies to identify the expired tokens by comparing the `validUntil` property with the current date and time.

Finally, it outputs the display name and validity expiration date for each expired token found for the respective application.

Please note that the script assumes you have the necessary permissions to access the Azure AD applications and query the Microsoft Graph API.